shallot-conformance-tier-split: S1
Split conformance.test.ts into a trimmed default-tier sentinel plus a by-path
tier sibling (conformance.tier.ts) holding the GPU pipeline-compiling arms.

The full 18-arm file ran 4.6–6.3 s straddling the 5000 ms per-file cap
(test-cap.ts), so the verdict was a host-load coin flip (4-in-6 red) with
zero conformance assertion failures — every red was the cap. The split is
the cap's own prescribed move (2): promote the slow arms to a by-path tier.

One roster in conformance-roster.ts, split by a declared predicate
(isPipelineCompiling: entries whose plugins include RenderPlugin). The
default-tier file keeps the cheap arms (Project, Mirror, Input,
Slab+Transforms, Orbit, Tween) plus the seeded non-idempotent red arm as
the sentinel the promotion law requires.
